backend/requirements.txt

fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
requests
celery[redis]
redis
python-dotenv
cryptography
plotly
